requirements.txt
opencv-python
numpy
trimesh
plyfile
pyyaml
scikit-image
scikit-learn
chumpy
tensorboard
kornia
loguru
pycocotools
yacs
lmdb
fire
setuptools
einops
tqdm
ipython
gym
transformers==4.44.2
OmegaConf
matplotlib
gradio==5.0.2
diffusers
rembg
segment_anything
jaxtyping
imageio
iopath
timm
open3d
pyrender
pytorch-lightning
smplx==0.1.28
chumpy @ git+https://github.com/zerchen/chumpy
xtcocotools
pandas
hydra-core
hydra-submitit-launcher
hydra-colorlog
pyrootutils
rich
webdataset
ultralytics
dill
